requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
psycopg2-binary~=2.9.3
django-cors-headers==4.3.1
djoser==2.2.2
django-filters==23.5